requirements.txt
openai-agents
python-dotenv
gradio